DyGLNet-main/requirements.txt
torch==2.4.1+cu121+cu121
torchvision==0.19.1+cu121+cu121
numpy==1.24.4
scipy==1.10.1
pandas==2.0.3
pillow==9.3.0
tqdm==4.64.1
matplotlib==3.6.2
scikit-learn==1.3.2